robot/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

google_api_core == 2.10.2
google_api_python_client == 2.66.0
google_auth_oauthlib == 0.7.1
